+++ test/System.Linq.Dynamic.Core.Tests/QueryableTests.Is,OfType,As,Cast.cs
@@ -488,6 +488,34 @@ namespace System.Linq.Dynamic.Core.Tests
             castDynamic.Should().BeEquivalentTo(new[] { 1, 2 });
         }
 
+        // #440
+        [Fact]
+        public void CastToIntUsingParentheses()
+        {
+            // Assign
+            var qry = new[]
+            {
+                new User
+                {
+                    Id = 1,
+                    DisplayName = "" + (char) 109
+                },
+                new User
+                {
+                    Id = 2,
+                    DisplayName = "abc"
+                }
+            }.AsQueryable();
+
+            // Act
+            var result1 = qry.Where("DisplayName.Any(int(it) >= 109) and Id > 0").ToDynamicArray<User>();
+            var result2 = qry.Where("Id > 0 && DisplayName.Any(int(it) >= 109)").ToDynamicArray<User>();
+
+            // Assert
+            result1.Should().HaveCount(1).And.Subject.First().Id.Should().Be(1);
+            result2.Should().HaveCount(1).And.Subject.First().Id.Should().Be(1);
+        }
+
         [Fact]
         public void CastToType_Dynamic_ActingOnIt()
         {